⚡️ Performance Improvements for Tooltips & Parameter Separators
React has an addiction to doing hydration as one long, unbroken task—so I've opted to explicitly defer some nonessential parts of the client to when the browser is less busy. Testing on my own machine, main thread blocking on problematic pages (e.g. `/skyrimse/script/po3_sksefunctions/`) went from about 40 seconds to around 1.5 seconds. Those 1.5 seconds were, by in large, spent on scheduling the deferred rendering/hydration of tooltips and parameter separators and should therefore be considered unrecoverable.

Additionally, I do not expect the particularly problematic pages to be navigated to directly, so they don't have any hydration steps to speak of. The loading indicator more than covers the rendering delays for typical navigation.
