+++ tsconfig.json
@@ -54,6 +54,7 @@
   ],
   "exclude": [
     "out",
-    "node_modules"
+    "node_modules",
+    "cache",
   ],
 }
